rename shared pkg to placement
